+++ src/styles/styledLoading.jsx
